tickets/src/events/listener/orderCreatedListener.ts
import { Listner, NotFoundError, OrderCreatedEvent, Subjects } from '@sj-bl/common'
import { Message } from 'node-nats-streaming';
import { Ticket } from '../../models/tickets';
import { TicketUpdatedPublisher } from '../publisher/ticketUpdatedPublisher';
import { queueGroupName } from './queueGroupName';



export class OrderCreatedListener extends Listner<OrderCreatedEvent>{
  subject: Subjects.OrderCreated = Subjects.OrderCreated;
  queueGroupName = queueGroupName;
  async onMessage(data: OrderCreatedEvent['data'], msg: Message) {

    const ticket = await Ticket.findById(data.ticket.id);

    if (!ticket) {
      throw new Error('Ticket Not Found');
    }

    ticket.set({ orderId: data.id });

    await ticket.save();

    new TicketUpdatedPublisher(this.client).publish({
      id: ticket.id,
      title: ticket.title,
      userId: ticket.userId,
      price: ticket.price,
      version: ticket.version,
      orderId: ticket.orderId
    })

    msg.ack()

  }
}